requirements.txt
PyYAML>=6.0,<7
requests>=2.31,<3
feedparser>=6.0.11,<7
beautifulsoup4>=4.12,<5
python-dateutil>=2.9,<3
Jinja2>=3.1,<4
jsonschema>=4.22,<5
trafilatura>=2,<3
openai>=1.40,<2
google-genai
